requirements/ci.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements/ci.txt requirements/base.txt requirements/test-tools.in
#
alabaster==0.7.12
amqp==2.5.2
babel==2.8.0
beautifulsoup4==4.8.2     # via webtest
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
coverage==4.5.4
cssselect==1.1.0          # via pyquery
django-admin-index==1.2.3
django-appconf==1.0.3
django-axes==5.2.2
django-camunda==0.4.0
django-choices==1.7.1
django-compat==1.0.15
django-docutils==0.4.0
django-filter==2.2.0
django-hijack-admin==2.1.10
django-hijack==2.1.10
django-ipware==2.1.0
django-jenkins==0.110.0
django-ordered-model==2.1.0
django-polymorphic==2.1.2
django-redis==4.11.0
django-rosetta==0.9.3
django-sniplates==0.6.0
django-solo==1.1.3
django-timeline-logger==1.1.2
django-webtest==1.9.7
django==2.2.10
djangorestframework==3.11.0
docutils==0.16
elastic-apm==5.3.2
factory-boy==2.12.0
faker==3.0.0              # via factory-boy
flower==0.9.3
freezegun==0.3.12
gemma-zds-client==0.13.1
idna==2.8
imagesize==1.2.0
inflection==0.3.1
jinja2==2.11.1
kombu==4.6.7
lxml==4.4.2               # via pyquery
markupsafe==1.1.1
packaging==20.1
pillow==7.0.0
polib==1.1.0
psycopg2==2.8.4
pygments==2.5.2
pyjwt==1.7.1
pyparsing==2.4.6
pyquery==1.4.1
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.10.3
pytz==2019.3
pyyaml==5.3
redis==3.3.11
requests-mock==1.7.0
requests==2.22.0
sentry-sdk==0.14.0
six==1.13.0
snowballstemmer==2.0.0
soupsieve==1.9.5          # via beautifulsoup4
sphinx==2.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
tblib==1.6.0
text-unidecode==1.3       # via faker
tornado==5.1.1
urllib3==1.25.7
uwsgi==2.0.18
vine==1.3.0
waitress==1.4.2           # via webtest
webob==1.8.5              # via webtest
webtest==2.0.33           # via django-webtest
wheel==0.29.0
zgw-consumers==0.5.1

# The following packages are considered to be unsafe in a requirements file:
# setuptools
